Moved the SLA metric classes and Amazon EC2 Template class to CloudSim Plus module, respectively to the packages slametrics and vmtemplates.
