+++ cloudsim-plus-testbeds/pom.xml
@@ -124,10 +124,5 @@
             <artifactId>json-simple</artifactId>
             <version>1.1.1</version>
         </dependency>
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.8.0</version>
-        </dependency>
     </dependencies>
 </project>
