+++ cloudsim-plus/src/main/java/org/cloudsimplus/slametrics/package-info.java
@@ -4,9 +4,10 @@
  * <a href="http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/viewing_metrics_with_cloudwatch.html">Amazon Cloudwatch</a>.
  *
  * <p>A JSON file can be read using a method such as the
- * {@link org.cloudsimplus.sla.metrics.SlaContract#getInstance(java.lang.String)}.</p>
+ * {@link org.cloudsimplus.slametrics.SlaContract#getInstance(java.lang.String)}.</p>
  *
  * @author raysaoliveira
- * @see org.cloudsimplus.sla.metrics.SlaContract
+ * @since CloudSim Plus 1.2.2
+ * @see org.cloudsimplus.slametrics.SlaContract
  */
-package org.cloudsimplus.sla.metrics;
+package org.cloudsimplus.slametrics;
